Implementation for a photo filter library and selector
Allows the editor screen to display a grid of buttons that apply
the given filter to the selected image. The filter can then be
used to process the image when uploading.

Lays the groundwork for building specialized filters based off
of CIFilter. Subclass BaseFilter and provide a name and override
the -processCIImage:withContext; to apply the effect to the image.

See #6